% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-my-class.R
\name{my_class}
\alias{my_class}
\title{Create your own class}
\usage{
my_class(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{args}
}
\value{
"my_class" object
}
\description{
Create your own class
}
